Support Inbox message functionality
The following operations require an approval by another employee:
-Price change - approved only by the CEO
-publish new version - approved by the managing editors.
An inbox message is sent to the approver which descides whether to
approve/decline the request. another inbox message is produced with
the approver's decision and is sent to the employee who made the
request. if approved- requested action is executed in addition.